// pronti-frontend-main/front/src/app/modulos/plan/form/form-plan.enum.ts
export enum FormPlanEnum {
  // nombre = 'nombre',

  modelo = 'modelo',
  precio = 'precio',
  inscripcion = 'inscripcion',
  cuotaMes12 = 'cuotaMes12',
  cuotaMes24 = 'cuotaMes24',
  cuotaMes36 = 'cuotaMes36',
  cuotaMes48 = 'cuotaMes48',
  cuotaMes60 = 'cuotaMes60',
  cuotaMes72 = 'cuotaMes72',
  cuotaMes84 = 'cuotaMes84'
}